pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "qq-court-claim-assistant-linux"
version = "0.1.0"
description = "QQ court claim assistant lightweight Linux deployment package"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "nonebot2[fastapi,websockets]>=2.4.0,<3.0.0",
  "nonebot-adapter-onebot>=2.4.6,<3.0.0",
  "httpx>=0.27.0,<1.0.0",
  "pydantic>=2.8.0,<3.0.0",
  "PyYAML>=6.0.1,<7.0.0",
  "python-dotenv>=1.0.1,<2.0.0",
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["bot_app*"]
